Move HTTPMethod and HTTPResponse to directives.py
